rename GreedyUtilityFunction to PredictedValueUtilityFunction
